Create CardGame Project
Expressions of CASE, HAVING, AND/OR clauses to interpret the data from card games among friends. Intent to show the highest possible points in a single game, the averages of which games were played and the results of a winning player based on the average game played with the highest possible points in a single game.